requirements.txt
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
config==0.5.1
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
imaplib2==3.6
iniconfig==2.0.0
install==1.3.5
outcome==1.3.0.post0
packaging==24.0
pluggy==1.5.0
PySocks==1.7.1
pytest==8.2.2
pytest-ordering==0.6
python-dotenv==1.0.1
requests==2.32.3
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
